requirements.txt
Flask==2.3.2
Pillow==10.0.1
torch==2.0.1
transformers==4.33.2
numpy==1.24.4
